backwards compatible secret renewal
