+++ aomi/render.py
@@ -10,45 +10,9 @@ from cryptorito import portable_b64decode, is_base64
 from aomi.helpers import merge_dicts, cli_hash, \
     path_pieces, abspath
 from aomi.template import render, load_vars
+from aomi.vault import renew_secret, is_aws
 import aomi.exceptions
 LOG = logging.getLogger(__name__)
-
-
-def grok_seconds(lease):
-    """Ensures that we are returning just seconds"""
-    if lease.endswith('s'):
-        return int(lease[0:-1])
-    elif lease.endswith('m'):
-        return int(lease[0:-1]) * 60
-    elif lease.endswith('h'):
-        return int(lease[0:-1]) * 3600
-
-    return None
-
-
-def is_aws(data):
-    """Takes a decent guess as to whether or not we are dealing with
-    an AWS secret blob"""
-    return 'access_key' in data and 'secret_key' in data
-
-
-def renew_secret(client, creds, opt):
-    """Renews a secret. This will occur unless the user has
-    specified on the command line that it is not neccesary"""
-    if opt.reuse_token:
-        return
-
-    seconds = grok_seconds(opt.lease)
-    if not seconds:
-        raise aomi.exceptions.AomiCommand("invalid lease %s" % opt.lease)
-
-    renew = client.renew_secret(creds['lease_id'], seconds)
-    # sometimes it takes a bit for vault to respond
-    # if we are within 5s then we are fine
-    if seconds - renew['lease_duration'] >= 5:
-        client.revoke_self_token()
-        e_msg = 'Unable to renew with desired lease'
-        raise aomi.exceptions.VaultConstraint(e_msg)
 
 
 def secret_key_name(path, key, opt):
